--- weapon/weaponIds.py
@@ -1,154 +1,0 @@
-weapon_IDs = {
-    # PHB
-    "Assault Cannon": "2yAiZeQHZbNXjqUs",
-    "Blaster Cannon": "atdw9BYBUaq3TBJc",
-    "Blaster Carbine": "ff3pk8eoISixRmoS",
-    "Blaster Pistol": "rsvtNAd7ynaSl9Vm",
-    "Blaster Rifle": "mtFnCiSUAj9kivGi",
-    "Bowcaster": "1SBIOPVaJBJOzMX3",
-    "Carbine Rifle": "ftt2CdiE87E6YH7o",
-    "Chaingun": "AMHhpY8JNWdgtTxt",
-    "Chakram": "SHFWAW9vX90h7wy2",
-    "Cycler Rifle": "EoDw5dGGjFL26UVF",
-    "Doubleblade": "FaI2hcLloxVCVhJs",
-    "Doublesaber": "k9itkDmC1x7Z1K7P",
-    "Doubleshoto": "5ITBFp9TFfWegA8a",
-    "Doublesword": "OAgCRzEXmHFrOrrQ",
-    "Greatsaber": "uH7eKtZhXWCTq28J",
-    "Heavy Bowcaster": "S2VEGzRNabUz9Y3v",
-    "Heavy Pistol": "xZmrMpjEoVEhiEFp",
-    "Heavy Repeater": "dnsnA7btkZWhauhS",
-    "Heavy Shotgun": "Rnnty5E35aSfX42x",
-    "Heavy Slugpistol": "6CKTZxpD5pCbDvTM",
-    "Hidden Blade": "G4XP83WKt0oks972",
-    "Hold-Out": "lMx7dwuE5IZtDPtz",
-    "Hunting Rifle": "HT9JkoXHsZ6X7W8q",
-    "Ion Carbine": "K9ejn3f6GGoY66yW",
-    "Ion Pistol": "jGDuLu25eqYklyqv",
-    "Ion Rifle": "oZYhBpB4G2pq02Rq",
-    "Light Pistol": "TMKqWa3KZKTxiaPB",
-    "Light Repeater": "N84CPvAv89ELqgJ7",
-    "Light Ring": "pIKgMdXvjQbTqMCx",
-    "Light Slugpistol": "J5eex5McFySyg66d",
-    "Lightclub": "uzeq8Hjm7oFw2rA7",
-    "Lightdagger": "jQc4ZmyZnEZXDZS7",
-    "Lightfoil": "tYdQWaOlSm97aYdw",
-    "Lightsaber": "HtN8KYXXMaJ9L94x",
-    "Lightsaber Pike": "ylA99WYUVY8U5z9U",
-    "Martial Lightsaber": "thGihwn0AfTo5Wu7",
-    "Needler": "EFCyA4xJQE25DEbw",
-    "Net": "j2g1D976omFXUQP1",
-    "Repeating Blaster": "zMxfSwEGBrTYbsP3",
-    "Revolver": "5N3UBYLZf785VxNK",
-    "Saberspear": "TWL4tclbh0dVt6HR",
-    "Saberwhip": "6VAWhI8EESUa217M",
-    "Scattergun": "AXba3svQTMscEGlf",
-    "Shotgun": "FN8cSlvwnzkIbbNa",
-    "Shotosaber": "zTg0qnvZJDAht3Zm",
-    "Slugpistol": "5xtJY3hAPoTGNhuD",
-    "Slugthrower": "W4G7FhgqrpFE0wdX",
-    "Sniper Rifle": "JliFzE6002igf46x",
-    "Subrepeater": "A7U4bqnbP9tFACPX",
-    "Techaxe": "G5bjOdE8vxCM8NOq",
-    "Techblade": "wwumzQvxwoPyyuTX",
-    "Techstaff": "I6UqxtQec0NXW8gv",
-    "Tranquilizer Rifle": "Lw6vUczorCLaGPFe",
-    "Unarmed Strike": "znMXEaqEAzwM6HSj",
-    "Vibroaxe": "8bcaFVQ91NKjrH2e",
-    "Vibrobaton": "X3EtJ3JgDWIPiED9",
-    "Vibroblade": "Axbve7PaSX9obY47",
-    "Vibrodagger": "XfUEDKVByHe83aSB",
-    "Vibrodart": "TM1rXOy8GZEbmP6x",
-    "Vibroknuckler": "omDKHum2Ybu2atpI",
-    "Vibrolance": "p5URddJqRu3grRQl",
-    "Vibromace": "jjulP09xLc8oYOZo",
-    "Vibropike": "JMj0t59NQtIU8Rbo",
-    "Vibrorapier": "YmIH2FjeQwjzNMFa",
-    "Vibrospear": "B0hPklxvpHwSxCMJ",
-    "Vibrostaff": "bNzwxj6dlIjwqTnH",
-    "Vibrosword": "OrWJNGr3WDulfOO0",
-    "Vibrowhip": "Rxa18dl0msDmUCB1",
-    "Wrist Launcher": "Jbf3fp7lslwSjWbW",
-    "Wristblaster": "Ax4RQZEjbsHcjfJf",
-
-    # EC
-    "Buster Saber": "ZF8JWL5b9beaqDDT",
-    "Cross-Saber": "FvJ9dzTZy4L3UPVo",
-    "Disguised Blade": "e6MVaUPHZskkCoWf",
-    "Hooked Vibroblade": "5gaBZ2WKdnVbQ6Rr",
-    "Jagged Vibroblade": "eN87BxUs8gQLNtbS",
-    "Riot Baton": "o89rRDni85TvThPk",
-    "Riot Shocker": "YbuEc7AYMjLZY27c",
-    "Saber Gauntlet": "dHLgqBFFDb6RPg2G",
-    "Shock Whip": "GTjR8Jeh7wbhWmqJ",
-    "Torpedo Launcher": "uXzS3fD2vuywQ4cd",
-    "Vibrobattleaxe": "ziwY4nKtv76cVmN7",
-    "Vibrobuster": "cTo7T8JMBxFwrGwi",
-    "Vibroclaymore": "v6jLemXa6gbz3Z35",
-    "Vibroglaive": "JQf8bC4u7zZmtMRF",
-
-    # WH
-    "Arc Caster": "S8avZvm86o2Nko2u",
-    "BKG": "T5aEHtouczCMtQBj",
-    "Bo-Rifle": "sb5fUN5pM4tZxCFi",
-    "Bolas": "c5CdA644mjKFQjvT",
-    "Bolt-Thrower": "ro9eKXF5HjQB73cV",
-    "Canesaber": "UvaJWx6zgw3U8Skb",
-    "Chained Dagger": "5Sp84fKN6xHQtR9Z",
-    "Chained Lightdagger": "W6Bt8mJtx6pnAXQh",
-    "Claymore Saber": "mXgmD7AwUtyDA7gg",
-    "Crossguard Saber": "aN6j3fQWNVoYDHZx",
-    "Disruptor Pistol": "Bcq4WouSfzTWqhWs",
-    "Disruptor Rifle": "Y7T9HviecBG9gd2n",
-    "Disruptorshiv": "3NuQW62wsn8CXL2d",
-    "Dual-Phase Saber": "mBc4rN3CJYVjKDRg",
-    "Echostaff": "LxYgJveBA3Fs5Vps",
-    "Electrobaton": "AHqQr9M78Hph8EuJ",
-    "Electrohammer": "3SG3QopLE6UMQxev",
-    "Electroprod": "5i4ob4YfRXMjHeEp",
-    "Electrostaff": "wRD2j6xhs94SU8CY",
-    "Electrovoulge": "JqSkXzfDQYn3Xmvr",
-    "Energy Bow": "hdNwKJKj8x4hW3vZ",
-    "Flechette Cannon": "LdG3iMiWDpu3vLgF",
-    "Grenade Launcher": "9Kix3A5unZKQDvCb",
-    "Guard Shoto": "XA7RB3AAREfa6BHy",
-    "Hand Blaster": "cNia5xdPtWzmEvbW",
-    "IWS": "3bBzRoEgTBpUGDzy",
-    "Incinerator Pistol": "SeRdVn2Vp7eRZJNk",
-    "Incinerator Sniper": "oKjN2m3VJT7xG8Kv",
-    "Lightaxe": "ZnbqnvcA33Csg5AS",
-    "Lightbow": "fNUuXGjqDL4ecq3Q",
-    "Lightfist": "t5Q6HfJahRPMaagq",
-    "Lightglaive": "aWZ2mXSdvyYKxKTN",
-    "Mancatcher": "8EnFjh5A2t3GJ3am",
-    "Nervebaton": "jhi3K2UhLy4Y92MZ",
-    "Nightstinger Rifle": "6DCjiUEs4aYiJDRz",
-    "Railgun": "XcSmwLUvYbESWJ9d",
-    "Retrosaber": "6aRVvBHyL9TrNPHc",
-    "Rocket Launcher": "EuMkteZuZoAAqS5q",
-    "Rotary Cannon": "toeCuDf27MV4mMZh",
-    "Saberaxe": "DDEs6KEvmjg8JjgP",
-    "Shatter Cannon": "5Nz8echLkBknCdHA",
-    "Shatter Pistol": "ayKYtWc7uf3X8Pd2",
-    "Shatter Rifle": "WctkrKLX4iQQ38o7",
-    "Shoulder Cannon": "9nWzXs6TfTPaPJ22",
-    "Sith Saber": "V9GZyE7sYjExoHXt",
-    "Sonic Pistol": "EC6kTdbScLKxjWDj",
-    "Sonic Rifle": "5vKpt5NMXQupUTjP",
-    "Switch Cannon": "kbecuy96r5Zqcu7S",
-    "Switch Carbine": "HPUkYoPSU2f48Jt5",
-    "Switch Pistol": "bR5q3bDxMGaBsema",
-    "Switch Rifle": "iVLFHv3UU9SqN6yQ",
-    "Switch Sniper": "D5RvahjHgYAy8exr",
-    "Vapor Projector": "JRztR5M8s6dGuPeH",
-    "Vibroclaw": "dAoHx3PAtPF5F7DB",
-    "Vibrocutter": "mCLffKzisJ6keNmj",
-    "Vibrohammer": "rf7AA4Lu4V2CZ4En",
-    "Vibroknife": "WP5EsyT2q5iNYSTr",
-    "Vibroshield": "wMi4S2zd7vhRE8ou",
-    "Vibrotonfa": "qkUM3yepBdKDzBWi",
-    "War Hat": "WSaaLfejMxSZd5x5",
-    "Warsword": "EULPzXe8REmARPGC",
-    "Wristblade": "BmSnTW8ocg3JzPJK",
-    "Wristsaber": "eg6zUB2qbPhXqKD7"
-};
