+++ weapons/generate.py
