+++ weapons/template.py
