+++ weapons/weapons_Ids.py
@@ -1,4 +1,4 @@
-weapons_Ids = {
+weapon_IDs = {
     # PHB
     "Assault Cannon": "2yAiZeQHZbNXjqUs",
     "Blaster Cannon": "atdw9BYBUaq3TBJc",
@@ -33,8 +33,8 @@ weapons_Ids = {
     "Lightclub": "uzeq8Hjm7oFw2rA7",
     "Lightdagger": "jQc4ZmyZnEZXDZS7",
     "Lightfoil": "tYdQWaOlSm97aYdw",
-    "Lightsaber Pike": "ylA99WYUVY8U5z9U",
     "Lightsaber": "HtN8KYXXMaJ9L94x",
+    "Lightsaber Pike": "ylA99WYUVY8U5z9U",
     "Martial Lightsaber": "thGihwn0AfTo5Wu7",
     "Needler": "EFCyA4xJQE25DEbw",
     "Net": "j2g1D976omFXUQP1",
@@ -69,5 +69,86 @@ weapons_Ids = {
     "Vibrosword": "OrWJNGr3WDulfOO0",
     "Vibrowhip": "Rxa18dl0msDmUCB1",
     "Wrist Launcher": "Jbf3fp7lslwSjWbW",
-    "Wristblaster": "Ax4RQZEjbsHcjfJf"
-}
+    "Wristblaster": "Ax4RQZEjbsHcjfJf",
+
+    # EC
+    "Buster Saber": "ZF8JWL5b9beaqDDT",
+    "Cross-Saber": "FvJ9dzTZy4L3UPVo",
+    "Disguised Blade": "e6MVaUPHZskkCoWf",
+    "Hooked Vibroblade": "5gaBZ2WKdnVbQ6Rr",
+    "Jagged Vibroblade": "eN87BxUs8gQLNtbS",
+    "Riot Baton": "o89rRDni85TvThPk",
+    "Riot Shocker": "YbuEc7AYMjLZY27c",
+    "Saber Gauntlet": "dHLgqBFFDb6RPg2G",
+    "Shock Whip": "GTjR8Jeh7wbhWmqJ",
+    "Torpedo Launcher": "uXzS3fD2vuywQ4cd",
+    "Vibrobattleaxe": "ziwY4nKtv76cVmN7",
+    "Vibrobuster": "cTo7T8JMBxFwrGwi",
+    "Vibroclaymore": "v6jLemXa6gbz3Z35",
+    "Vibroglaive": "JQf8bC4u7zZmtMRF",
+
+    # WH
+    "Arc Caster": "S8avZvm86o2Nko2u",
+    "BKG": "T5aEHtouczCMtQBj",
+    "Bo-Rifle": "sb5fUN5pM4tZxCFi",
+    "Bolas": "c5CdA644mjKFQjvT",
+    "Bolt-Thrower": "ro9eKXF5HjQB73cV",
+    "Canesaber": "UvaJWx6zgw3U8Skb",
+    "Chained Dagger": "5Sp84fKN6xHQtR9Z",
+    "Chained Lightdagger": "W6Bt8mJtx6pnAXQh",
+    "Claymore Saber": "mXgmD7AwUtyDA7gg",
+    "Crossguard Saber": "aN6j3fQWNVoYDHZx",
+    "Disruptor Pistol": "Bcq4WouSfzTWqhWs",
+    "Disruptor Rifle": "Y7T9HviecBG9gd2n",
+    "Disruptorshiv": "3NuQW62wsn8CXL2d",
+    "Dual-Phase Saber": "mBc4rN3CJYVjKDRg",
+    "Echostaff": "LxYgJveBA3Fs5Vps",
+    "Electrobaton": "AHqQr9M78Hph8EuJ",
+    "Electrohammer": "3SG3QopLE6UMQxev",
+    "Electroprod": "5i4ob4YfRXMjHeEp",
+    "Electrostaff": "wRD2j6xhs94SU8CY",
+    "Electrovoulge": "JqSkXzfDQYn3Xmvr",
+    "Energy Bow": "hdNwKJKj8x4hW3vZ",
+    "Flechette Cannon": "LdG3iMiWDpu3vLgF",
+    "Grenade Launcher": "9Kix3A5unZKQDvCb",
+    "Guard Shoto": "XA7RB3AAREfa6BHy",
+    "Hand Blaster": "cNia5xdPtWzmEvbW",
+    "IWS": "3bBzRoEgTBpUGDzy",
+    "Incinerator Pistol": "SeRdVn2Vp7eRZJNk",
+    "Incinerator Sniper": "oKjN2m3VJT7xG8Kv",
+    "Lightaxe": "ZnbqnvcA33Csg5AS",
+    "Lightbow": "fNUuXGjqDL4ecq3Q",
+    "Lightfist": "t5Q6HfJahRPMaagq",
+    "Lightglaive": "aWZ2mXSdvyYKxKTN",
+    "Mancatcher": "8EnFjh5A2t3GJ3am",
+    "Nervebaton": "jhi3K2UhLy4Y92MZ",
+    "Nightstinger Rifle": "6DCjiUEs4aYiJDRz",
+    "Railgun": "XcSmwLUvYbESWJ9d",
+    "Retrosaber": "6aRVvBHyL9TrNPHc",
+    "Rocket Launcher": "EuMkteZuZoAAqS5q",
+    "Rotary Cannon": "toeCuDf27MV4mMZh",
+    "Saberaxe": "DDEs6KEvmjg8JjgP",
+    "Shatter Cannon": "5Nz8echLkBknCdHA",
+    "Shatter Pistol": "ayKYtWc7uf3X8Pd2",
+    "Shatter Rifle": "WctkrKLX4iQQ38o7",
+    "Shoulder Cannon": "9nWzXs6TfTPaPJ22",
+    "Sith Saber": "V9GZyE7sYjExoHXt",
+    "Sonic Pistol": "EC6kTdbScLKxjWDj",
+    "Sonic Rifle": "5vKpt5NMXQupUTjP",
+    "Switch Cannon": "kbecuy96r5Zqcu7S",
+    "Switch Carbine": "HPUkYoPSU2f48Jt5",
+    "Switch Pistol": "bR5q3bDxMGaBsema",
+    "Switch Rifle": "iVLFHv3UU9SqN6yQ",
+    "Switch Sniper": "D5RvahjHgYAy8exr",
+    "Vapor Projector": "JRztR5M8s6dGuPeH",
+    "Vibroclaw": "dAoHx3PAtPF5F7DB",
+    "Vibrocutter": "mCLffKzisJ6keNmj",
+    "Vibrohammer": "rf7AA4Lu4V2CZ4En",
+    "Vibroknife": "WP5EsyT2q5iNYSTr",
+    "Vibroshield": "wMi4S2zd7vhRE8ou",
+    "Vibrotonfa": "qkUM3yepBdKDzBWi",
+    "War Hat": "WSaaLfejMxSZd5x5",
+    "Warsword": "EULPzXe8REmARPGC",
+    "Wristblade": "BmSnTW8ocg3JzPJK",
+    "Wristsaber": "eg6zUB2qbPhXqKD7"
+};
